% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythymbox-export.R
\name{read_xspf}
\alias{read_xspf}
\title{read XSPF playlist}
\usage{
read_xspf(x, ...)
}
\arguments{
\item{x}{path to XSPF playlist}
}
\description{
read XSPF playlist
}
\examples{
pg <- read_xspf("~/Desktop/Car Music.xspf")
}
