% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythymbox-export.R
\name{sync}
\alias{sync}
\title{sync files to phone}
\usage{
sync(files, dir = "~/Desktop/Phone_Music/", ...)
}
\arguments{
\item{files}{character vector of file paths}

\item{dir}{directory to put the files}
}
\description{
sync files to phone
}
\examples{

pg <- read_xspf("~/Desktop/Car Music.xspf")
sync(pg)
}
